Rename `arch` to `Arch`
