+++ kernel/Arch/i686/asm/interrupt.S
