+++ kernel/Arch/i686/boot/stivale2.S
