+++ kernel/Arch/i686/gdt.cpp
@@ -10,7 +10,7 @@
  * @copyright Copyright the Xyris Contributors (c) 2019
  *
  */
-#include <arch/i686/gdt.hpp>
+#include <Arch/i686/gdt.hpp>
 #include <lib/string.hpp>
 
 // Defined in the gdt_flush.s file.
