+++ kernel/Arch/i686/idt.hpp
@@ -10,7 +10,7 @@
  */
 
 #pragma once
-#include <arch/i686/Arch.i686.hpp>
+#include <Arch/i686/Arch.i686.hpp>
 #include <stdint.h>
 
 /* Segment selectors */
