+++ kernel/Arch/i686/linker.ld
