+++ kernel/Arch/i686/timer.hpp
@@ -9,7 +9,7 @@
  *
  */
 #pragma once
-#include <arch/i686/Arch.i686.hpp>
+#include <Arch/i686/Arch.i686.hpp>
 #include <stdint.h>
 
 #define TIMER_COMMAND_PORT 0x43
