+++ kernel/Arch/i686/tss.hpp
@@ -9,7 +9,7 @@
  *
  */
 #pragma once
-#include <arch/i686/Arch.i686.hpp>
+#include <Arch/i686/Arch.i686.hpp>
 #include <stdint.h>
 
 /**
