+++ kernel/dev/spkr/spkr.cpp
@@ -9,9 +9,9 @@
  *
  */
 
-#include <arch/Arch.hpp>
+#include <Arch/Arch.hpp>
 #include <dev/spkr/spkr.hpp>
-#include <arch/i686/timer.hpp> // TODO: Remove ASAP
+#include <Arch/i686/timer.hpp> // TODO: Remove ASAP
 
 static void spkr_tone(uint32_t freq) {
     uint32_t div;
