+++ kernel/sys/trace.cpp
@@ -9,7 +9,7 @@
  *
  */
 
-#include <arch/Arch.hpp>
+#include <Arch/Arch.hpp>
 #include <dev/graphics/console.hpp>
 #include <dev/serial/rs232.hpp>
 #include <lib/stdio.hpp>
